tests: Changed demo assemblies to a test assembly for integration test
