+++ scripts/demo.bat
@@ -4,24 +4,24 @@ SET workspace=%~dp0..\
 @REM set filterregex=--filterregex \.file\= --filterregex \.attributes\.guid\= --filterregex numberOfImplementedInterfaces --filterregex implementedInterfaces
 set filterregex=
 
-SET net60dll=%workspace%src\dscom.demo\assembly1\bin\Release\net6.0\dSPACE.Runtime.InteropServices.DemoAssembly1.dll
-SET net48dll=%workspace%src\dscom.demo\assembly1\bin\Release\net48\dSPACE.Runtime.InteropServices.DemoAssembly1.dll
+SET net60dll=%workspace%src\dscom.test.assembly\bin\Release\net8.0\dSPACE.Runtime.InteropServices.Test.Assembly.dll
+SET net48dll=%workspace%src\dscom.test.assembly\bin\Release\net48\dSPACE.Runtime.InteropServices.Test.Assembly.dll
 
-del %workspace%src\dscom.demo\assembly1\bin\Release\net48\*.tlb 
-del %workspace%src\dscom.demo\assembly1\bin\Release\net6.0\*.tlb 
-del %workspace%src\dscom.demo\assembly1\bin\Release\net6.0\*.yaml
-del %workspace%src\dscom.demo\assembly1\bin\Release\net48\*.yaml
+del %workspace%src\dscom.test.assembly\bin\Release\net48\*.tlb 
+del %workspace%src\dscom.test.assembly\bin\Release\net8.0\*.tlb 
+del %workspace%src\dscom.test.assembly\bin\Release\net8.0\*.yaml
+del %workspace%src\dscom.test.assembly\bin\Release\net48\*.yaml
 
 dotnet build -c Release "%workspace%dscom.sln"
 IF ERRORLEVEL 1 goto error
 
 @REM dscom
 echo ############## dscom.exe tlbexport
-%workspace%src\dscom.client\bin\Release\net6.0\dscom.exe tlbexport /verbose "%net60dll%" "/out:%net60dll%.tlb"
+%workspace%src\dscom.client\bin\Release\net8.0\dscom.exe tlbexport /verbose "%net60dll%" "/out:%net60dll%.tlb"
 IF ERRORLEVEL 1 goto error
 
 echo ############## dscom.exe tlbdump
-%workspace%src\dscom.client\bin\Release\net6.0\dscom.exe tlbdump %filterregex% "/tlbrefpath:%net60dll%.tlb/.." "%net60dll%.tlb" "/out:%net60dll%.yaml"
+%workspace%src\dscom.client\bin\Release\net8.0\dscom.exe tlbdump %filterregex% "/tlbrefpath:%net60dll%.tlb/.." "%net60dll%.tlb" "/out:%net60dll%.yaml"
 IF ERRORLEVEL 1 goto error
 
 WHERE tlbexp
@@ -43,7 +43,7 @@ tlbexp /win64 /verbose "%net48dll%" "/out:%net48dll%.tlb"
 IF ERRORLEVEL 1 goto error
 
 echo ############## dscom.exe tlbdump
-%workspace%src\dscom.client\bin\Release\net6.0\dscom.exe tlbdump %filterregex% "/tlbrefpath:%net48dll%.tlb/.." "%net48dll%.tlb" "/out:%net48dll%.yaml"
+%workspace%src\dscom.client\bin\Release\net8.0\dscom.exe tlbdump %filterregex% "/tlbrefpath:%net48dll%.tlb/.." "%net48dll%.tlb" "/out:%net48dll%.yaml"
 IF ERRORLEVEL 1 goto error
 
 WHERE code
