+++ scripts/test.coverage.bat
@@ -7,5 +7,5 @@ IF NOT EXIST %TEST_RESULTS% (
     mkdir %TEST_RESULTS%
 )
 
-dotnet test -f net6.0 -c Release /p:CollectCoverage=true /p:CoverletOutputFormat=lcov /p:CoverletOutput=%TEST_RESULTS%\lcov.info --logger "trx;LogFileName=%TEST_RESULTS%\results.xml" %PWD%..\dscom.sln 
-COPY %TEST_RESULTS%\lcov.net6.0.info %TEST_RESULTS%\lcov.info
+dotnet test -f net8.0 -c Release /p:CollectCoverage=true /p:CoverletOutputFormat=lcov /p:CoverletOutput=%TEST_RESULTS%\lcov.info --logger "trx;LogFileName=%TEST_RESULTS%\results.xml" %PWD%..\dscom.sln 
+COPY %TEST_RESULTS%\lcov.net8.0.info %TEST_RESULTS%\lcov.info
