+++ scripts/test.coverage.report.bat
@@ -9,5 +9,5 @@ IF NOT EXIST %TEST_RESULTS% (
 
 dotnet tool install -g dotnet-reportgenerator-globaltool
 
-dotnet test  -f net6.0 -c Release /p:CollectCoverage=true /p:CoverletOutputFormat=cobertura /p:CoverletOutput="%TEST_RESULTS%\\"  %PWD%..\dscom.sln 
-reportgenerator -reports:"%TEST_RESULTS%\coverage.net6.0.cobertura.xml" -targetdir:"%TEST_RESULTS%\report" -reporttypes:Html
+dotnet test  -f net8.0 -c Release /p:CollectCoverage=true /p:CoverletOutputFormat=cobertura /p:CoverletOutput="%TEST_RESULTS%\\"  %PWD%..\dscom.sln 
+reportgenerator -reports:"%TEST_RESULTS%\coverage.net8.0.cobertura.xml" -targetdir:"%TEST_RESULTS%\report" -reporttypes:Html
