--- src/dscom.demo/assembly1/DemoEnums.cs
@@ -1,36 +1,0 @@
-// Copyright 2022 dSPACE GmbH, Mark Lechtermann, Matthias Nissen and Contributors
-// 
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-// 
-//     http://www.apache.org/licenses/LICENSE-2.0
-// 
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-using System.Runtime.InteropServices;
-using dSPACE.Runtime.InteropServices.DemoAssembly2;
-
-namespace dSPACE.Runtime.InteropServices.DemoAssembly1.EnumNamespace1
-{
-    [ComVisible(Constants.DEFAULT_VISIBILITY)]
-    public enum DemoEnum
-    {
-        A,
-        B
-    }
-}
-
-namespace dSPACE.Runtime.InteropServices.DemoAssembly1.EnumNamespace2
-{
-    [ComVisible(Constants.DEFAULT_VISIBILITY)]
-    public enum DemoEnum
-    {
-        A,
-        B
-    }
-}
