--- src/dscom.demo/assembly4/assembly4.csproj
@@ -1,23 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <TargetFrameworks>net8.0;net6.0;net48</TargetFrameworks>
-    <PlatformTarget>AnyCPU</PlatformTarget>
-    <LangVersion>10.0</LangVersion>
-    <ImplicitUsings>enable</ImplicitUsings>
-    <Nullable>enable</Nullable>
-    <RootNamespace>dSPACE.Runtime.InteropServices.DemoAssembly4</RootNamespace>
-    <AssemblyName>dSPACE.Runtime.InteropServices.DemoAssembly4</AssemblyName>
-    <EnforceCodeStyleInBuild>true</EnforceCodeStyleInBuild>
-    <AnalysisMode>Recommended</AnalysisMode>
-    <EnableNETAnalyzers>true</EnableNETAnalyzers>
-    <Platforms>AnyCPU;x86</Platforms>
-    <EnableComHosting>true</EnableComHosting>
-    <NoWarn>NU5104,CS1591</NoWarn>    
-  </PropertyGroup>
-
-  <ItemGroup>
-    <ProjectReference Include="..\..\dscom\dscom.csproj" />
-  </ItemGroup>
-
-</Project>
