--- src/dscom.demo/assembly5/ExplodeyBaseClass.cs
@@ -1,11 +1,0 @@
-using System.Windows.Controls;
-
-namespace dSPACE.Runtime.InteropServices.DemoAssembly5;
-
-public class ExplodeyBaseClass : UserControl
-{
-    private sealed class PrivateImplementationDetails
-    {
-        // required to reproduce #259
-    }
-}
