--- src/dscom.demo/assembly5/ExplodeyMember.cs
@@ -1,9 +1,0 @@
-using System.Windows.Controls;
-
-namespace dSPACE.Runtime.InteropServices.DemoAssembly5;
-
-public class ExplodeyMember
-{
-    [System.Diagnostics.CodeAnalysis.SuppressMessage("Performance", "CA1822:Mark members as static", Justification = "Example")]
-    public UserControl Control => null!;
-}
