--- src/dscom.demo/assembly5/ExplodeyParameter.cs
@@ -1,13 +1,0 @@
-using System.Windows.Controls;
-
-namespace dSPACE.Runtime.InteropServices.DemoAssembly5;
-
-public class ExplodeyParameter
-{
-    [System.Diagnostics.CodeAnalysis.SuppressMessage("Performance", "CA1822:Mark members as static", Justification = "example")]
-    [System.Diagnostics.CodeAnalysis.SuppressMessage("Style", "IDE0060:Remove unused parameter", Justification = "example")]
-    public void Example(UserControl thing)
-    {
-        // nop
-    }
-}
