--- src/dscom.demo/assembly5/assembly5.csproj
@@ -1,26 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <TargetFrameworks>net8.0-windows;net6.0-windows;net48</TargetFrameworks>
-    <PlatformTarget>AnyCPU</PlatformTarget>
-    <LangVersion>10.0</LangVersion>
-    <ImplicitUsings>enable</ImplicitUsings>
-    <Nullable>enable</Nullable>
-    <RootNamespace>dSPACE.Runtime.InteropServices.DemoAssembly5</RootNamespace>
-    <AssemblyName>dSPACE.Runtime.InteropServices.DemoAssembly5</AssemblyName>
-    <EnforceCodeStyleInBuild>true</EnforceCodeStyleInBuild>
-    <AnalysisMode>Recommended</AnalysisMode>
-    <EnableNETAnalyzers>true</EnableNETAnalyzers>
-    <Platforms>AnyCPU;x86</Platforms>
-    <EnableComHosting>true</EnableComHosting>
-    <NoWarn>NU5104,CS1591</NoWarn>
-  </PropertyGroup>
-  <PropertyGroup>
-    <UseWPF>true</UseWPF>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <ProjectReference Include="..\..\dscom\dscom.csproj" />
-  </ItemGroup>
-
-</Project>
