--- src/dscom.demo/assembly6/assembly6.csproj
@@ -1,23 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <TargetFrameworks>net8.0</TargetFrameworks>
-    <PlatformTarget>AnyCPU</PlatformTarget>
-    <LangVersion>10.0</LangVersion>
-    <ImplicitUsings>enable</ImplicitUsings>
-    <Nullable>enable</Nullable>
-    <RootNamespace>dSPACE.Runtime.InteropServices.DemoAssembly6</RootNamespace>
-    <AssemblyName>dSPACE.Runtime.InteropServices.DemoAssembly6</AssemblyName>
-    <EnforceCodeStyleInBuild>true</EnforceCodeStyleInBuild>
-    <AnalysisMode>Recommended</AnalysisMode>
-    <EnableNETAnalyzers>true</EnableNETAnalyzers>
-    <Platforms>AnyCPU;x86</Platforms>
-    <EnableComHosting>true</EnableComHosting>
-    <NoWarn>NU5104,CS1591</NoWarn>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <ProjectReference Include="..\..\dscom\dscom.csproj" />
-  </ItemGroup>
-
-</Project>
