+++ src/dscom.test.assembly/DemoHiddenAndRestricted.cs
@@ -15,7 +15,7 @@
 using System.Runtime.InteropServices;
 using dSPACE.Runtime.InteropServices.Attributes;
 
-namespace dSPACE.Runtime.InteropServices.DemoAssembly4;
+namespace dSPACE.Runtime.InteropServices.Test;
 
 [ComVisible(true)]
 public interface IDemoInterfaceVisible
